moved separated components and pages in front end files
