+++ hiring_agency_frontend/src/App.js
@@ -1,50 +1,50 @@
 import './App.css';
 import 'animate.css';
 import 'bootstrap/dist/css/bootstrap.css';
-import Login from './components/login';
-import Home from './components/home';
+import Login from './pages/login';
+import Home from './pages/home';
 import NavMenu from './components/navMenu';
-import ChangePassword from './components/changePassword';
-import SetSecurityQuestions from './components/setSecurityQuestions';
-import AddStaff from './components/addStaff';
+import ChangePassword from './pages/changePassword';
+import SetSecurityQuestions from './pages/setSecurityQuestions';
+import AddStaff from './pages/addStaff';
 import { HashRouter as Router, Routes, Route } from 'react-router-dom';
-import AccountLocked from './components/accountLocked';
+import AccountLocked from './pages/accountLocked';
 import { useEffect, useState } from 'react';
 import { getAuthToken, isUserLoggedIn } from './utils/storage';
 import { getUser } from './api/authRequests';
-import CreateAdvertisement from './components/createAdvertisement';
+import CreateAdvertisement from './pages/createAdvertisement';
 
-import ViewAdvertisement from './components/viewAdvertisement';
-import CareTakerAccountRequest from './components/careTakerAccountRequest';
-import CareTakerAccountRequestSuccess from './components/careTakerAccountRequestSuccess';
-import CreateApplication from './components/createApplication.jsx';
-import CreateApplicationSuccess from './components/createApplicationSuccess';
-import PendingCareTakerRequests from './components/pendingCareTakerRequests';
-import CareTakerAccountRequestDetail from './components/careTakerAccountRequestDetail';
-import CareTakerAccountRequestApproved from './components/careTakerAccountRequestApproved';
-import CareTakerAccountRequestRejected from './components/careTakerAccountRequestRejected';
+import ViewAdvertisement from './pages/viewAdvertisement';
+import CareTakerAccountRequest from './pages/careTakerAccountRequest';
+import CareTakerAccountRequestSuccess from './pages/careTakerAccountRequestSuccess';
+import CreateApplication from './pages/createApplication.jsx';
+import CreateApplicationSuccess from './pages/createApplicationSuccess';
+import PendingCareTakerRequests from './pages/pendingCareTakerRequests';
+import CareTakerAccountRequestDetail from './pages/careTakerAccountRequestDetail';
+import CareTakerAccountRequestApproved from './pages/careTakerAccountRequestApproved';
+import CareTakerAccountRequestRejected from './pages/careTakerAccountRequestRejected';
 import StaffRoute from './routes/staffRoute';
 import AdminRoute from './routes/adminRoute';
 import UnauthenticatedRoute from './routes/unauthenticatedRoute';
 import CareTakerRoute from './routes/careTakerRoute';
-import CreateServiceRequest from './components/createServiceRequest';
-import CreateServiceRequestSuccess from './components/createServiceRequestSuccess';
-import ServiceRequests from './components/serviceRequests';
-import CreateJobSuccess from './components/createAdvertisementSuccess';
-import ServiceRequestDetail from './components/serviceRequestDetail';
-import JobRequestList from './components/jobRequestList';
-import JobRequesDetail from './components/jobRequestDetail';
+import CreateServiceRequest from './pages/createServiceRequest';
+import CreateServiceRequestSuccess from './pages/createServiceRequestSuccess';
+import ServiceRequests from './pages/serviceRequests';
+import CreateJobSuccess from './pages/createAdvertisementSuccess';
+import ServiceRequestDetail from './pages/serviceRequestDetail';
+import JobRequestList from './pages/jobRequestList';
+import JobRequesDetail from './pages/jobRequestDetail';
 import JobRequesReject from './components/jobRequestRejected';
-import JobRequesApproved from './components/jobRequestApproved';
-import StaffList from './components/staffList';
-import StaffDetail from './components/staffDetail';
-import StaffStatusChanged from './components/staffStatusChanged';
-import CareTakerList from './components/careTakerList';
-import CareTakerDetail from './components/careTakerDetail';
-import CareTakerStatusChanged from './components/careTakerStatusChanged';
-import MyMenu from './components/myMenu';
-import BillingAccounts from './components/billingAccounts';
-import BillingAccountDetail from './components/billingAccountDetail';
+import JobRequesApproved from './pages/jobRequestApproved';
+import StaffList from './pages/staffList';
+import StaffDetail from './pages/staffDetail';
+import StaffStatusChanged from './pages/staffStatusChanged';
+import CareTakerList from './pages/careTakerList';
+import CareTakerDetail from './pages/careTakerDetail';
+import CareTakerStatusChanged from './pages/careTakerStatusChanged';
+import MyMenu from './pages/myMenu';
+import BillingAccounts from './pages/billingAccounts';
+import BillingAccountDetail from './pages/billingAccountDetail';
 
 
 function App() {
