+++ hiring_agency_frontend/src/pages/createAdvertisementSuccess.jsx
