+++ hiring_agency_frontend/src/pages/createApplicationSuccess.jsx
